just added explanations for semantic and syntax errors

+++ src/Part_1/Method.java
@@ -13,24 +13,23 @@ import Exceptions.BadRatingException;
 import Exceptions.BadScoreException;
 import Exceptions.BadGenreException;
 
-
 public class Method {
 
-    public static void partitionMovies(String inputFileName, FileWriter badRecordsWriter, FileWriter manifestWriter){
-       
+    public static void partitionMovies(String inputFileName, FileWriter badRecordsWriter, FileWriter manifestWriter) {
+
         try (BufferedReader reader = new BufferedReader(new FileReader(inputFileName))) {
             String line;
-       
+
             while ((line = reader.readLine()) != null) {
                 String[] movieData = line.split(",");
-       
+
                 if (validateAndWriteRecord(movieData)) {
-       
+
                     String genre = movieData[3].trim(); // Assuming genre is at index 3
                     FileWriter genreWriter = new FileWriter(genre + ".csv");
                     genreWriter.write(line + "\n");
                     genreWriter.close();
-       
+
                 } else {
                     badRecordsWriter.write(line + "\n");
                 }
@@ -41,48 +40,60 @@ public class Method {
             manifestWriter.write(manifestEntry + "\n");
         } catch (IOException e) {
             System.err.println("Error processing input file: " + e.getMessage());
-        } catch(PartOneException | MissingFieldsException | MissingQuotesException | ExcessFieldsException |
-            BadYearException | BadDurationException | BadRatingException | BadScoreException |
-            BadGenreException exception){
+        } catch (PartOneException | MissingFieldsException | MissingQuotesException | ExcessFieldsException
+                | BadYearException | BadDurationException | BadRatingException | BadScoreException
+                | BadGenreException exception) {
             System.err.println("Error processing input file: " + exception.getMessage());
         }
     }
 
-    public static boolean validateAndWriteRecord(String[] movieData) throws IOException, PartOneException, MissingFieldsException, MissingQuotesException, ExcessFieldsException, BadYearException, BadDurationException, BadRatingException, BadScoreException, BadGenreException{
-        // Syntax check
+    public static boolean validateAndWriteRecord(String[] movieData)
+            throws IOException, PartOneException, MissingFieldsException, MissingQuotesException, ExcessFieldsException,
+            BadYearException, BadDurationException, BadRatingException, BadScoreException, BadGenreException {
+        // ----- Syntax check -----
+        // checks if movies have the required 10 fields
         if (movieData.length != 10) {
             throw new MissingFieldsException();
         }
 
+        // checks if titles have missing quotes
         if (!movieData[1].startsWith("\"") || !movieData[1].endsWith("\"")) {
             throw new MissingQuotesException();
         }
 
+        // checks if titles have excess fields
         if (movieData[1].split("\"").length > 3) {
             throw new ExcessFieldsException();
         }
 
-        // Semantic check
+        // ----- Semantic check -----
+        // checks if the year is between 1990 and 1999
         int year = Integer.parseInt(movieData[0]);
         if (year < 1990 || year > 1999) {
             throw new BadYearException();
         }
 
+        // checks if the durations are between 0 and 300 minutes
         int duration = Integer.parseInt(movieData[2]);
         if (duration < 0 || duration > 300) {
             throw new BadDurationException();
         }
 
-        int score = Integer.parseInt(movieData[5]);
+        // checks if the scores are between 0 and 10
+        double score = Double.parseDouble(movieData[5]);
         if (score < 0 || score > 10) {
             throw new BadScoreException();
         }
 
+        // checks if the ratings are the valid ones (PG, Unrated, G, R, PG-13, NC-17)
         String rating = movieData[4];
         if (!isValidRating(rating)) {
             throw new BadRatingException();
         }
 
+        // checks if the genres are the valid ones (musical, comedy, animation,
+        // adventure, drama, crime, biography, horror, action, documentary, fantasy,
+        // mystery, sci-fi, family, romance, thriller, western)
         String genre = movieData[3];
         if (!isValidGenre(genre)) {
             throw new BadGenreException();
@@ -97,9 +108,9 @@ public class Method {
     }
 
     private static boolean isValidGenre(String genre) {
-        String[] validGenres = {"musical", "comedy", "animation", "adventure", "drama", "crime",
+        String[] validGenres = { "musical", "comedy", "animation", "adventure", "drama", "crime",
                 "biography", "horror", "action", "documentary", "fantasy", "mystery", "sci-fi",
-                "family", "romance", "thriller", "western"};
+                "family", "romance", "thriller", "western" };
         for (String validGenre : validGenres) {
             if (genre.equals(validGenre)) {
                 return true;
